requirements.txt
aiogram==3.17.0
aiohttp==3.11.11
aiosqlite==0.20.0
python-dotenv==1.0.1
loguru==0.7.3